+++ cmd/resim/commands/experience.go
@@ -102,10 +102,12 @@ const (
 	experienceKey                     = "experience"
 	experiencesConfigKey              = "experiences-config"
 	experiencesCloneKey               = "clone"
+	experiencesSyncNoArchiveKey       = "no-archive"
 	experiencesUpdateConfigKey        = "update-config"
 	experienceIDKey                   = "id"
 	experienceDescriptionKey          = "description"
 	experienceLocationKey             = "location"
+	experienceLocationsKey            = "locations"
 	experienceLaunchProfileKey        = "launch-profile"
 	experienceGithubKey               = "github"
 	experienceTagKey                  = "tag"
@@ -123,7 +125,8 @@ func init() {
 	createExperienceCmd.Flags().String(experienceDescriptionKey, "", "The description of the experience")
 	createExperienceCmd.MarkFlagRequired(experienceDescriptionKey)
 	createExperienceCmd.Flags().StringSlice(experienceLocationKey, []string{}, "The location(s) of the experience, comma separated if more than one. e.g. an S3 URI for the experience folder")
-	createExperienceCmd.MarkFlagRequired(experienceLocationKey)
+	createExperienceCmd.Flags().MarkDeprecated(experienceLocationKey, "please use --locations flag instead (e.g. --locations s3://my-bucket/my-folder,s3://my-bucket/my-other-folder)")
+	createExperienceCmd.Flags().StringSlice(experienceLocationsKey, []string{}, "The locations (e.g. S3 URIs) of the experience, comma separated if more than one")
 	createExperienceCmd.Flags().String(experienceLaunchProfileKey, "", "The UUID of the launch profile for this experience")
 	createExperienceCmd.Flags().MarkDeprecated(experienceLaunchProfileKey, "launch profiles are deprecated in favor of systems to define resource requirements")
 	createExperienceCmd.Flags().Bool(experienceGithubKey, false, "Whether to output format in github action friendly format")
@@ -162,7 +165,8 @@ func init() {
 	updateExperienceCmd.MarkFlagRequired(experienceKey)
 	updateExperienceCmd.Flags().String(experienceNameKey, "", "New value for the name of the experience")
 	updateExperienceCmd.Flags().String(experienceDescriptionKey, "", "New value for the description of the experience")
-	updateExperienceCmd.Flags().String(experienceLocationKey, "", "New value for the location of the experience, e.g. an S3 URI for the experience folder")
+	updateExperienceCmd.Flags().String(experienceLocationKey, "", "New value for the location (e.g. S3 URI) of the experience (deprecated, use --locations instead)")
+	updateExperienceCmd.Flags().MarkDeprecated(experienceLocationKey, "please use --locations flag instead")
 	updateExperienceCmd.Flags().StringSlice(experienceSystemsKey, []string{}, "A list of system names or IDs to register as compatible with the experience")
 	updateExperienceCmd.Flags().StringSlice(experienceTagsKey, []string{}, "A list of experience tag names or IDs to apply to the experience")
 	updateExperienceCmd.Flags().Duration(experienceTimeoutKey, 1*time.Hour, "The timeout for the experience container. Default is 1 hour. Please use GoLang duration format e.g. 1h, 1m, 1s, etc.")
@@ -198,8 +202,12 @@ func init() {
 	syncExperienceCmd.Flags().String(experiencesConfigKey, "", "The path of the experiences config file to sync")
 	syncExperienceCmd.MarkFlagRequired(experiencesConfigKey)
 	syncExperienceCmd.Flags().Bool(experiencesUpdateConfigKey, false, "Whether to update the passed-in config in-place")
+	syncExperienceCmd.Flags().Bool(experiencesSyncNoArchiveKey, false, "Whether to archive experiences not listed in the config file")
+
 	syncExperienceCmd.Flags().Bool(experiencesCloneKey, false, "Whether to clone the existing database state to the config file rather than the other way around")
 	syncExperienceCmd.MarkFlagsMutuallyExclusive(experiencesUpdateConfigKey, experiencesCloneKey)
+	syncExperienceCmd.MarkFlagsMutuallyExclusive(experiencesSyncNoArchiveKey, experiencesCloneKey)
+
 	experienceCmd.AddCommand(syncExperienceCmd)
 
 	// Systems-related sub-commands:
@@ -239,7 +247,14 @@ func createExperience(ccmd *cobra.Command, args []string) {
 		log.Fatal("empty experience description")
 	}
 
-	experienceLocations := viper.GetStringSlice(experienceLocationKey)
+	var experienceLocations []string
+	// Prioritize the new locations flag over the deprecated location flag
+	if viper.IsSet(experienceLocationsKey) {
+		experienceLocations = viper.GetStringSlice(experienceLocationsKey)
+	} else if viper.IsSet(experienceLocationKey) {
+		experienceLocations = viper.GetStringSlice(experienceLocationKey)
+	}
+
 	if len(experienceLocations) == 0 {
 		log.Fatal("empty experience locations")
 	}
@@ -336,10 +351,10 @@ func createExperience(ccmd *cobra.Command, args []string) {
 		fmt.Println("Created experience successfully!")
 		fmt.Printf("Experience ID: %s\n", experience.ExperienceID.String())
 		if isCloud && objectsCount != nil && *objectsCount > 0 {
-			fmt.Printf("ReSim found %v file(s) in experience location:\n", *objectsCount)
+			fmt.Printf("ReSim found %v file(s) in experience locations:\n", *objectsCount)
 			OutputJson(*objectsInExperience)
 		} else {
-			fmt.Println("WARNING: ReSim could not find any files in the provided location.")
+			fmt.Println("WARNING: ReSim could not find any files in the provided locations.")
 		}
 	}
 }
@@ -400,7 +415,11 @@ func updateExperience(ccmd *cobra.Command, args []string) {
 		updateExperienceInput.Experience.Description = Ptr(viper.GetString(experienceDescriptionKey))
 		updateMask = append(updateMask, "description")
 	}
-	if viper.IsSet(experienceLocationKey) {
+	if viper.IsSet(experienceLocationsKey) {
+		locations := viper.GetStringSlice(experienceLocationsKey)
+		updateExperienceInput.Experience.Locations = &locations
+		updateMask = append(updateMask, "locations")
+	} else if viper.IsSet(experienceLocationKey) {
 		updateExperienceInput.Experience.Location = Ptr(viper.GetString(experienceLocationKey))
 		updateMask = append(updateMask, "location")
 	}
@@ -674,10 +693,11 @@ func syncExperience(ccmd *cobra.Command, args []string) {
 	projectID := getProjectID(Client, viper.GetString(experienceProjectKey))
 	configPath := viper.GetString(experiencesConfigKey)
 	updateConfig := viper.GetBool(experiencesUpdateConfigKey)
+	shouldArchive := !viper.GetBool(experiencesSyncNoArchiveKey)
 	clone := viper.GetBool(experiencesCloneKey)
 
 	if !clone {
-		experience_sync.SyncExperiences(Client, projectID, configPath, updateConfig)
+		experience_sync.SyncExperiences(Client, projectID, configPath, updateConfig, shouldArchive)
 	} else {
 		experience_sync.CloneExperiences(Client, projectID, configPath)
 	}