+++ cmd/resim/commands/sync/experiences_test.go
@@ -4,15 +4,18 @@ import (
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 	"gopkg.in/yaml.v3"
-	"log"
 	"slices"
 	"testing"
 )
 
 // A helper so we can load currentStateData and configData from yaml for convenience.
 func loaderHelper(t *testing.T, currentStateData string, configData string, currentTags []string, currentSystems []string) (DatabaseState, ExperienceSyncConfig) {
-	var currentExperiences []*Experience
+	var currentExperiences []Experience
 	err := yaml.Unmarshal([]byte(currentStateData), &currentExperiences)
+	assert.NoError(t, err, "failed to unmarshal YAML")
+	err = NormalizeExperiences(currentExperiences)
+	assert.NoError(t, err, "invalid experiences")
+
 	assert.NoError(t, err, "failed to unmarshal YAML")
 	currentState := DatabaseState{
 		ExperiencesByName: map[string]*Experience{},
@@ -36,19 +39,22 @@ func loaderHelper(t *testing.T, currentStateData string, configData string, curr
 
 	}
 	// Populate the tag and system sets based off what's listed in the YAML snippet.
-	for _, exp := range currentExperiences {
+	for ii := range currentExperiences {
+		exp := &currentExperiences[ii]
 		currentState.ExperiencesByName[exp.Name] = exp
 		for _, tag := range exp.Tags {
-			currentState.TagSetsByName[tag].ExperienceIDs[exp.ExperienceID.ID] = struct{}{}
+			currentState.TagSetsByName[tag].ExperienceIDs[*exp.ExperienceID] = struct{}{}
 		}
 		for _, system := range exp.Systems {
-			currentState.SystemSetsByName[system].ExperienceIDs[exp.ExperienceID.ID] = struct{}{}
+			currentState.SystemSetsByName[system].ExperienceIDs[*exp.ExperienceID] = struct{}{}
 		}
 	}
 
 	var config ExperienceSyncConfig
 	err = yaml.Unmarshal([]byte(configData), &config)
 	assert.NoError(t, err, "failed to unmarshal YAML")
+	err = NormalizeExperiences(config.Experiences)
+	assert.NoError(t, err, "invalid experiences")
 
 	return currentState, config
 }
@@ -59,7 +65,7 @@ func TestUpdateExperiencesEmpty(t *testing.T) {
 	currentState := DatabaseState{}
 
 	// ACTION
-	experienceUpdates, err := computeExperienceUpdates(&config, currentState)
+	experienceUpdates, err := computeExperienceUpdates(&config, currentState, true)
 
 	// VERIFICATION
 	assert.NoError(t, err)
@@ -77,16 +83,18 @@ experiences:
     description: This is a test experience
     locations:
       - s3://my-favorite-bucket/foo
-    environment_variables:
+    environmentVariables:
       - name: ENV_VAR_1
         value: value1
-    cache_exempt: true
-    container_timeout_seconds: 7200
+    cacheExempt: true
+    containerTimeoutSeconds: 7200
+managedExperienceTags:
+managedTestSuites:
 `
 	currentState, config := loaderHelper(t, currentStateData, configData, nil, nil)
 
 	// ACTION
-	experienceUpdates, err := computeExperienceUpdates(&config, currentState)
+	experienceUpdates, err := computeExperienceUpdates(&config, currentState, true)
 
 	// VERIFICATION
 	assert.NoError(t, err)
@@ -96,7 +104,7 @@ experiences:
 
 	match, exists := experienceUpdates.MatchedExperiencesByNewName[config.Experiences[0].Name]
 	assert.True(t, exists, "Expected experience in updates")
-	assert.Same(t, config.Experiences[0], match.New, "Should be the same object (pointer equality)")
+	assert.Same(t, &config.Experiences[0], match.New, "Should be the same object (pointer equality)")
 	assert.Nil(t, match.Original, "Experience is not new")
 }
 
@@ -105,14 +113,14 @@ func TestArchiveSingleExperience(t *testing.T) {
 	currentStateData := `
   - name: Test Experience
     description: This is a test experience
-    experience_id: "3dd91177-1e66-426c-bf5b-fb46fe4a0c3b"
+    experienceID: "3dd91177-1e66-426c-bf5b-fb46fe4a0c3b"
     locations:
       - s3://my-favorite-bucket/foo
-    environment_variables:
+    environmentVariables:
       - name: ENV_VAR_1
         value: value1
-    cache_exempt: true
-    container_timeout_seconds: 7200
+    cacheExempt: true
+    containerTimeoutSeconds: 7200
 `
 	configData := `
 experiences:
@@ -120,7 +128,7 @@ experiences:
 	currentState, config := loaderHelper(t, currentStateData, configData, nil, nil)
 
 	// ACTION
-	experienceUpdates, err := computeExperienceUpdates(&config, currentState)
+	experienceUpdates, err := computeExperienceUpdates(&config, currentState, true)
 
 	// VERIFICATION
 	assert.NoError(t, err)
@@ -131,9 +139,6 @@ experiences:
 	match, exists := experienceUpdates.MatchedExperiencesByNewName["Test Experience"]
 	assert.True(t, exists, "Expected experience in updates")
 	assert.Same(t, currentState.ExperiencesByName["Test Experience"], match.Original, "Should be the same object (pointer equality)")
-
-	log.Print(match.Original.Locations)
-	log.Print(match.New.Locations)
 
 	assert.Equal(t, match.Original.Name, match.New.Name)
 	assert.Equal(t, match.Original.Description, match.New.Description)
@@ -152,25 +157,25 @@ func TestUpdateSingleExperiencesByNameAndID(t *testing.T) {
 	currentStateData := `
   - name: experience-to-update-by-name
     description: This is a test experience
-    experience_id: "3dd91177-1e66-426c-bf5b-fb46fe4a0c3b"
+    experienceID: "3dd91177-1e66-426c-bf5b-fb46fe4a0c3b"
     locations:
       - s3://my-favorite-bucket/foo
-    environment_variables:
+    environmentVariables:
       - name: ENV_VAR_1
         value: value1
-    cache_exempt: true
-    container_timeout_seconds: 7200
-    archived: true  # This will be restured
+    cacheExempt: true
+    containerTimeoutSeconds: 7200
+    archived: true  # This will be returned
   - name: experience-to-update-by-id
     description: This is a test experience2
-    experience_id: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    experienceID: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
     locations:
       - s3://my-favorite-bucket/bar
-    environment_variables:
+    environmentVariables:
       - name: ENV_VAR_2
         value: value2
-    cache_exempt: false
-    container_timeout_seconds: 7200
+    cacheExempt: false
+    containerTimeoutSeconds: 7200
 `
 	configData := `
 experiences:
@@ -178,26 +183,26 @@ experiences:
     description: This is my new experience"
     locations:
       - s3://my-favorite-bucket/my-new-location
-    environment_variables:
+    environmentVariables:
       - name: ENV_VAR_1
         value: value_new
-    cache_exempt: false
-    container_timeout_seconds: 7300
+    cacheExempt: false
+    containerTimeoutSeconds: 7300
   - name: new-name-for-experience-to-update-by-id
     description: This is a test experience2
-    experience_id: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    experienceID: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
     locations:
       - s3://my-favorite-bucket/bar
-    environment_variables:
+    environmentVariables:
       - name: ENV_VAR_2
         value: value2
-    cache_exempt: false
-    container_timeout_seconds: 7300
+    cacheExempt: false
+    containerTimeoutSeconds: 7300
 `
 	currentState, config := loaderHelper(t, currentStateData, configData, nil, nil)
 
 	// ACTION
-	experienceUpdates, err := computeExperienceUpdates(&config, currentState)
+	experienceUpdates, err := computeExperienceUpdates(&config, currentState, true)
 
 	// VERIFICATION
 	assert.NoError(t, err)
@@ -211,7 +216,7 @@ experiences:
 		match, exists := experienceUpdates.MatchedExperiencesByNewName[config.Experiences[i].Name]
 		assert.True(t, exists, "Expected experience in updates")
 		assert.Same(t, currentState.ExperiencesByName[originalNames[i]], match.Original, "Should be the same object (pointer equality)")
-		assert.Same(t, config.Experiences[i], match.New, "Should be the same object (pointer equality)")
+		assert.Same(t, &config.Experiences[i], match.New, "Should be the same object (pointer equality)")
 	}
 }
 
@@ -222,35 +227,47 @@ func TestFailsOnAmbiguousRenaming(t *testing.T) {
 	// indicates it's the update for current-experience. This is a failure.
 	currentStateData := `
   - name: current-experience
-    experience_id: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    description: Some experience
+    locations: ["somewhere_over_the_rainbow"]
+    experienceID: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
 `
 	configData := `
 experiences:
   - name: current-experience
+    description: Some experience
+    locations: ["somewhere_over_the_rainbow"]
   - name: new-name-for-current-experience
-    experience_id: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    description: Some experience
+    locations: ["somewhere_over_the_rainbow"]
+    experienceID: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
 `
 	currentState, config := loaderHelper(t, currentStateData, configData, nil, nil)
 
 	// ACTION / VERIFICATION
-	_, err := computeExperienceUpdates(&config, currentState)
+	_, err := computeExperienceUpdates(&config, currentState, true)
 	assert.Error(t, err)
 
 	// SETUP
 	currentStateData = `
   - name: current-experience
-    experience_id: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
+    experienceID: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
 `
 	configData = `
 experiences:
   - name: new-name-for-current-experience
-    experience_id: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
+    experienceID: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
   - name: current-experience
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
 `
 	currentState, config = loaderHelper(t, currentStateData, configData, nil, nil)
 
 	// ACTION / VERIFICATION
-	_, err = computeExperienceUpdates(&config, currentState)
+	_, err = computeExperienceUpdates(&config, currentState, true)
 	assert.Error(t, err)
 }
 
@@ -258,17 +275,21 @@ func TestFailsOnNonExistentID(t *testing.T) {
 	// SETUP
 	currentStateData := `
   - name: current-experience
-    experience_id: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    experienceID: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
 `
 	configData := `
 experiences:
   - name: new-name-for-current-experience
-    experience_id: "8f8e2af7-28d4-4462-8025-d313ccb61bd2"
+    experienceID: "8f8e2af7-28d4-4462-8025-d313ccb61bd2"
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
 `
 	currentState, config := loaderHelper(t, currentStateData, configData, nil, nil)
 
 	// ACTION / VERIFICATION
-	_, err := computeExperienceUpdates(&config, currentState)
+	_, err := computeExperienceUpdates(&config, currentState, true)
 	assert.Error(t, err)
 }
 
@@ -276,19 +297,25 @@ func TestFailsOnDuplicateID(t *testing.T) {
 	// SETUP
 	currentStateData := `
   - name: current-experience
-    experience_id: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    experienceID: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
 `
 	configData := `
 experiences:
   - name: new-name-for-current-experience
-    experience_id: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    experienceID: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
   - name: other-new-name-for-current-experience
-    experience_id: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    experienceID: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
 `
 	currentState, config := loaderHelper(t, currentStateData, configData, nil, nil)
 
 	// ACTION / VERIFICATION
-	_, err := computeExperienceUpdates(&config, currentState)
+	_, err := computeExperienceUpdates(&config, currentState, true)
 	assert.Error(t, err)
 }
 
@@ -301,21 +328,29 @@ func TestFailsOnClobberingExistingWithRename(t *testing.T) {
 	// don't support this.
 	currentStateData := `
   - name: new-name
-    experience_id: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    experienceID: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
   - name: old-name
-    experience_id: "8f8e2af7-28d4-4462-8025-d313ccb61bd2"
+    experienceID: "8f8e2af7-28d4-4462-8025-d313ccb61bd2"
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
 `
 	configData := `
 experiences:
   - name: new-name
-    experience_id: "8f8e2af7-28d4-4462-8025-d313ccb61bd2"
+    experienceID: "8f8e2af7-28d4-4462-8025-d313ccb61bd2"
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
   - name: old-name
-    experience_id: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    experienceID: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
 `
 	currentState, config := loaderHelper(t, currentStateData, configData, nil, nil)
 
 	// ACTION / VERIFICATION
-	_, err := computeExperienceUpdates(&config, currentState)
+	_, err := computeExperienceUpdates(&config, currentState, true)
 	assert.Error(t, err)
 }
 
@@ -323,19 +358,27 @@ func TestFailsOnNameCollision(t *testing.T) {
 	// SETUP
 	currentStateData := `
   - name: new-name
-    experience_id: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    experienceID: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
   - name: old-name
-    experience_id: "8f8e2af7-28d4-4462-8025-d313ccb61bd2"
+    experienceID: "8f8e2af7-28d4-4462-8025-d313ccb61bd2"
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
 `
 	configData := `
 experiences:
   - name: new-name
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
   - name: new-name
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
 `
 	currentState, config := loaderHelper(t, currentStateData, configData, nil, nil)
 
 	// ACTION / VERIFICATION
-	_, err := computeExperienceUpdates(&config, currentState)
+	_, err := computeExperienceUpdates(&config, currentState, true)
 	assert.Error(t, err)
 }
 
@@ -343,31 +386,43 @@ func TestAddRemoveTags(t *testing.T) {
 	// SETUP
 	currentStateData := `
   - name: Test Experience
-    experience_id: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    experienceID: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
     tags: []
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
   - name: Unrenamed Experience
-    experience_id: "62501c04-3da2-4a46-94b1-ab90e32b2059"
+    experienceID: "62501c04-3da2-4a46-94b1-ab90e32b2059"
     tags: []
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
   - name: Old regression experience
-    experience_id: "cddca442-9c25-4c06-9023-a4edfe9258a3"
+    experienceID: "cddca442-9c25-4c06-9023-a4edfe9258a3"
     tags: ["regression", "my-special-tag"]
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
 `
 	configData := `
-managed_experience_tags:
+managedExperienceTags:
   - regression
 experiences:
   - name: Test Experience
     tags: ["regression"]
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
   - name: Renamed Experience
-    experience_id: "62501c04-3da2-4a46-94b1-ab90e32b2059"
+    experienceID: "62501c04-3da2-4a46-94b1-ab90e32b2059"
     tags: ["regression", "my-special-tag"]
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
   - name: Old regression experience
     tags: []
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
 `
 	currentState, config := loaderHelper(t, currentStateData, configData, []string{"regression", "my-special-tag"}, nil)
 
 	// ACTION
-	experienceUpdates, err := computeExperienceUpdates(&config, currentState)
+	experienceUpdates, err := computeExperienceUpdates(&config, currentState, true)
 
 	// VERIFICATION
 	assert.NoError(t, err)
@@ -375,20 +430,20 @@ experiences:
 	tagUpdates, contains := experienceUpdates.TagUpdatesByName["regression"]
 	assert.True(t, contains, "Tag 'regression' should be contained in TagUpdatesByName")
 
-	addedDesiredExperience := slices.Contains(tagUpdates.Additions, config.Experiences[0])
+	addedDesiredExperience := slices.Contains(tagUpdates.Additions, &config.Experiences[0])
 	assert.True(t, addedDesiredExperience, "Not going to tag desired experience")
-	addedDesiredExperience = slices.Contains(tagUpdates.Additions, config.Experiences[1])
+	addedDesiredExperience = slices.Contains(tagUpdates.Additions, &config.Experiences[1])
 	assert.True(t, addedDesiredExperience, "Not going to tag desired experience")
-	removedDesiredExperience := slices.Contains(tagUpdates.Removals, config.Experiences[2])
+	removedDesiredExperience := slices.Contains(tagUpdates.Removals, &config.Experiences[2])
 	assert.True(t, removedDesiredExperience, "Not going to untag desired experience")
 
 	tagUpdates, contains = experienceUpdates.TagUpdatesByName["my-special-tag"]
 	assert.True(t, contains, "Tag 'my-special-tag' should be contained in TagUpdatesByName")
-	addedDesiredExperience = slices.Contains(tagUpdates.Additions, config.Experiences[1])
+	addedDesiredExperience = slices.Contains(tagUpdates.Additions, &config.Experiences[1])
 	assert.True(t, addedDesiredExperience, "Not going to tag desired experience")
 
 	// We should *NOT* remove the unmanaged "my-special-tag" tag
-	removedDesiredExperience = slices.Contains(tagUpdates.Removals, config.Experiences[2])
+	removedDesiredExperience = slices.Contains(tagUpdates.Removals, &config.Experiences[2])
 	assert.False(t, removedDesiredExperience, "Going to untag desired experience with unmanaged tag.")
 }
 
@@ -396,21 +451,23 @@ func TestAddSystems(t *testing.T) {
 	// SETUP
 	currentStateData := `
   - name: Test Experience
-    experience_id: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    experienceID: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
     systems: []
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
 `
 	configData := `
-managed_experience_systems:
-  - regression
 experiences:
   - name: Test Experience But with New Name
-    experience_id: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    experienceID: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
     systems: ["planner"]
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
 `
 	currentState, config := loaderHelper(t, currentStateData, configData, nil, []string{"planner"})
 
 	// ACTION
-	experienceUpdates, err := computeExperienceUpdates(&config, currentState)
+	experienceUpdates, err := computeExperienceUpdates(&config, currentState, true)
 
 	// VERIFICATION
 	assert.NoError(t, err)
@@ -418,7 +475,7 @@ experiences:
 	systemUpdates, contains := experienceUpdates.SystemUpdatesByName["planner"]
 	assert.True(t, contains, "System 'planner' should be contained in SystemUpdatesByName")
 
-	addedDesiredExperience := slices.Contains(systemUpdates.Additions, config.Experiences[0])
+	addedDesiredExperience := slices.Contains(systemUpdates.Additions, &config.Experiences[0])
 	assert.True(t, addedDesiredExperience, "Not going to add system desired experience")
 }
 
@@ -426,18 +483,20 @@ func TestReviseTestSuite(t *testing.T) {
 	// SETUP
 	currentStateData := `
   - name: Test Experience
-    experience_id: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    experienceID: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
 `
 	configData := `
-managed_experience_systems:
-  - regression
-managed_test_suites:
+managedTestSuites:
   - name: "Nightly CI"
     experiences:
      - Test Experience But with New Name
 experiences:
   - name: Test Experience But with New Name
-    experience_id: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    experienceID: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
 `
 	currentState, config := loaderHelper(t, currentStateData, configData, nil, nil)
 	testSuiteID := uuid.New()
@@ -446,7 +505,7 @@ experiences:
 	}
 
 	// ACTION
-	experienceUpdates, err := computeExperienceUpdates(&config, currentState)
+	experienceUpdates, err := computeExperienceUpdates(&config, currentState, true)
 
 	// VERIFICATION
 	assert.NoError(t, err)
@@ -455,25 +514,29 @@ experiences:
 	assert.Equal(t, update.Name, "Nightly CI")
 	assert.Equal(t, update.TestSuiteID, testSuiteID)
 	assert.Len(t, update.Experiences, 1)
-	assert.Equal(t, update.Experiences[0], config.Experiences[0])
+	assert.Equal(t, update.Experiences[0], &config.Experiences[0])
 }
 
 func TestReviseTestSuiteFailOnOldName(t *testing.T) {
 	// SETUP
 	currentStateData := `
   - name: Test Experience Old Name
-    experience_id: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    experienceID: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
 `
 	configData := `
-managed_experience_systems:
+managedExperienceTags:
   - regression
-managed_test_suites:
+managedTestSuites:
   - name: "Nightly CI"
     experiences:
      - Test Experience Old Name
 experiences:
   - name: Test Experience But with New Name
-    experience_id: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    experienceID: "628eccf2-2621-4fdf-a8d8-c6b057ce2f0d"
+    description: Some current experience
+    locations: ["somewhere_over_the_rainbow"]
 `
 	currentState, config := loaderHelper(t, currentStateData, configData, nil, nil)
 	testSuiteID := uuid.New()
@@ -482,8 +545,52 @@ experiences:
 	}
 
 	// ACTION
-	_, err := computeExperienceUpdates(&config, currentState)
+	_, err := computeExperienceUpdates(&config, currentState, true)
 
 	// VERIFICATION
 	assert.Error(t, err)
+}
+
+func TestNoArchiveSingleExperience(t *testing.T) {
+	// SETUP
+	currentStateData := `
+  - name: Test Experience
+    description: This is a test experience
+    experienceID: "3dd91177-1e66-426c-bf5b-fb46fe4a0c3b"
+    locations:
+      - s3://my-favorite-bucket/foo
+    environmentVariables:
+      - name: ENV_VAR_1
+        value: value1
+    cacheExempt: true
+    containerTimeoutSeconds: 7200
+`
+	configData := `
+experiences:
+`
+	currentState, config := loaderHelper(t, currentStateData, configData, nil, nil)
+
+	// ACTION
+	shouldArchive := false
+	experienceUpdates, err := computeExperienceUpdates(&config, currentState, shouldArchive)
+
+	// VERIFICATION
+	assert.NoError(t, err)
+
+	assert.Len(t, experienceUpdates.MatchedExperiencesByNewName, 1, "Should be one experience update")
+	assert.Empty(t, experienceUpdates.TagUpdatesByName, "Should be no tag updates")
+
+	match, exists := experienceUpdates.MatchedExperiencesByNewName["Test Experience"]
+	assert.True(t, exists, "Expected experience in updates")
+	assert.Same(t, currentState.ExperiencesByName["Test Experience"], match.Original, "Should be the same object (pointer equality)")
+
+	assert.Equal(t, match.Original.Name, match.New.Name)
+	assert.Equal(t, match.Original.Description, match.New.Description)
+	assert.Equal(t, match.Original.Name, match.New.Name)
+	assert.Equal(t, match.Original.Profile, match.New.Profile)
+	assert.Equal(t, match.Original.ExperienceID, match.New.ExperienceID)
+	assert.Equal(t, match.Original.EnvironmentVariables, match.New.EnvironmentVariables)
+	assert.Equal(t, match.Original.CacheExempt, match.New.CacheExempt)
+	assert.Equal(t, match.Original.ContainerTimeoutSeconds, match.New.ContainerTimeoutSeconds)
+	assert.False(t, match.New.Archived, "Experience should not be archived")
 }